GIFConverter/flet/requirements.txt
pillow
flet